Split some task-related tests into a separate file
